requirements.txt
openai
python-dotenv
fastapi
streamlit
uvicorn
langchain-community
langchain-openai
tabulate
sqlalchemy
langchain
pandas
openai